requirements/base.txt
Django
Twisted[tls,http2]
attrs
celery
channels
daphne
dj-rest-auth
django-admin-autocomplete-filter
django-admin-env-notice
django-cors-headers
django-filter
django-health-check
django-model-utils
django-postgres-extra
django-redis
django-sql-dashboard
djangorestframework
djangorestframework-simplejwt
drf-nested-routers
httpx
orjson
pandas
psycopg2-binary
rich